+++ src/apis/rest/client.js
@@ -1,0 +1,10 @@
+import axios from 'axios';
+
+const Client = {
+  path: {
+    bithumbPublciApi: process.env.REACT_APP_BITHUMB_PUBLIC_API_PATH,
+  },
+  credentialsInstance: axios.create({ withCredentials: true }),
+  publicInstance: axios.create(),
+};
+export default Client;
