+++ src/apis/rest/index.js
@@ -1,10 +1,1 @@
-import axios from 'axios';
-
-const Client = {
-  path: {
-    bithumbPublciApi: process.env.REACT_APP_BITHUMB_PUBLIC_API_PATH,
-  },
-  credentialsInstance: axios.create({ withCredentials: true }),
-  publicInstance: axios.create(),
-};
-export default Client;
+export { default as bithumbServices } from "@apis/rest/bithumbServices";
