--- src/apis/rest/services/index.js
@@ -1,1 +1,0 @@
-export { default as bithumbServices } from "@apis/rest/services/bithumbServices";
